fitur 14: pengiriman delivery

+++ c02_5/src/main/java/com/ta/c02_5/repository/DeliveryDB.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 
 public interface DeliveryDB extends JpaRepository<DeliveryModel, Long> {
     List<DeliveryModel> findByOrderByPegawaiAsc();
-    Optional<DeliveryModel> findByIdDelivery(Long idDelivery);
+    Optional<DeliveryModel> findByIdDelivery(Integer idDelivery);
     Optional<DeliveryModel> findByPegawai(Long pegawai);
 }
